+++ tags/src/main/java/com/thousandeyes/sdk/tags/model/Tag.java
@@ -59,7 +59,7 @@ public class Tag {
   private AccessType accessType;
 
   public static final String JSON_PROPERTY_AID = "aid";
-  private Integer aid;
+  private Long aid;
 
   public static final String JSON_PROPERTY_COLOR = "color";
   private String color;
@@ -97,7 +97,7 @@ public class Tag {
   @JsonCreator
   public Tag(
     @JsonProperty(JSON_PROPERTY_ASSIGNMENTS) List<Assignment> assignments, 
-    @JsonProperty(JSON_PROPERTY_AID) Integer aid, 
+    @JsonProperty(JSON_PROPERTY_AID) Long aid, 
     @JsonProperty(JSON_PROPERTY_CREATE_DATE) String createDate, 
     @JsonProperty(JSON_PROPERTY_ID) UUID id, 
     @JsonProperty(JSON_PROPERTY_LEGACY_ID) BigDecimal legacyId
@@ -158,7 +158,7 @@ public class Tag {
   @JsonProperty(JSON_PROPERTY_AID)
   @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
 
-  public Integer getAid() {
+  public Long getAid() {
     return aid;
   }
